requirements.txt
django
djangorestframework
pandas
django-jet
